apps/api/requirements.txt
# FastAPI + Server
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9
websockets>=12.0

# Database
sqlalchemy[asyncio]>=2.0.35
asyncpg>=0.29.0
alembic>=1.13.0
psycopg[binary]>=3.2.0

# Auth
pyjwt>=2.9.0
cryptography>=43.0.0

# HTTP Client
httpx>=0.27.0
tenacity>=9.0.0

# LLM Clients
anthropic>=0.39.0
openai>=1.51.0

# SQL Parsing (for Databricks module)
sqlglot>=25.0.0

# Confluence
atlassian-python-api>=3.41.0
markdownify>=0.13.0

# Pydantic
pydantic>=2.9.0
pydantic-settings>=2.5.0

# Async I/O
aiofiles>=24.1.0

# Utilities
python-dotenv>=1.0.0
